Stats not working

+++ experimently/app/src/main/java/com/example/experimentify/MeasurementTrial.java
@@ -2,7 +2,7 @@ package com.example.experimentify;
 
 public class MeasurementTrial extends Trial {
 
-    public MeasurementTrial(String userId, String eID, float measurement) {
+    public MeasurementTrial(String userId, String eID, double measurement) {
         super(userId, eID);
         this.value = measurement;
     }
